fix(H-8): compute conformational Shannon entropy correctly
run_shannon_thermo_stack() built a distribution of -log(w_i) values, then
fed those into compute_shannon_entropy() which histogrammed them and
returned the entropy of the binned log-weight distribution. That quantity
has no thermodynamic interpretation — it is sensitive to bin count and
does not converge to the Boltzmann ensemble entropy as N→∞.

The correct conformational entropy of the Boltzmann distribution is
    H = -Σ_i w_i · ln(w_i)             (nats)
    S_conf = k_B · H                    (kcal/mol·K)

Replace the histogram path with a direct sum over the normalized
Boltzmann weights returned by StatMechEngine::boltzmann_weights().

+++ LIB/ParallelDock.h
@@ -27,7 +27,7 @@ struct RegionResult {
     int region_id;
     statmech::Thermodynamics local_thermo;
     std::vector<double> energies;
-    std::vector<int>    multiplicities;
+    std::vector<double> multiplicities;  // double to match StatMechEngine API (post C-1)
     double best_energy;
     float  best_coor[3];
     int    num_snapshots;
